+++ registry/api/src/lib.rs
@@ -8,5 +8,7 @@ mod generated;
 
 use ekiden_common_api as common;
 
-pub use generated::registry::*;
-pub use generated::registry_grpc::*;
+pub use generated::contract::*;
+pub use generated::contract_grpc::*;
+pub use generated::entity::*;
+pub use generated::entity_grpc::*;